+++ packages/sdk-shared/src/test-setup.js
@@ -150,6 +150,25 @@ const DEVICES = {
 }
 
 const BROWSERS = {
+  'ff': {
+    type: 'sauce',
+    url: SAUCE_SERVER_URL,
+    capabilities: {
+      w3c: {
+        'browserName': 'firefox',
+        'platformName': 'Windows 10',
+        'browserVersion': '72.0',
+      },
+      legacy: {
+        'browserName': 'firefox',
+        'platform': 'Windows 10',
+        'version': '62.0',
+      },
+    },
+    options: {
+      ...SAUCE_CREDENTIALS,
+    },
+  },
   'edge-18': {
     type: 'sauce',
     url: SAUCE_SERVER_URL,